Fourt Commit

+++ DemoWebApiService/Controllers/OrderController.cs
@@ -25,32 +25,45 @@ namespace DemoWebApiService.Controllers
             _orderRepository = new OrderRepository(context);
             _orderItemRepository = new OrderItemRepository(context);
         }
-        [HttpPost("api/Order/PostOrder/{_order}")]
-        public ActionResult PostOrder(Order _order)
+        [HttpPost("api/Order/PostOrder/{orderParam}")]
+        public ActionResult PostOrder(Order orderParam)
         {
             try
-            { 
-            //Order table
-            if (_order.OrderID ==0)
-                _orderRepository.AddOrder(_mapper.Map<Demo.Core.Data.Model.Order>(_order));
-            else
-                _orderRepository.UpdateOrder(_mapper.Map<Demo.Core.Data.Model.Order>(_order));
-
-            //OrderItems table
-            foreach (var item in _order.OrderItems)
             {
-                if (item.OrderItemID == 0)
-                    _orderItemRepository.AddOrderItem(_mapper.Map<Demo.Core.Data.Model.OrderItem>(item));
-                else
-                    _orderItemRepository.UpdateOrderItem(_mapper.Map<Demo.Core.Data.Model.OrderItem>(item));
-            }
-                // Delete for OrderItems
-                foreach (var id in _order.DeletedOrderItemIDs.Split(',').Where(x => x != ""))
+                //Order table
+                if (orderParam.OrderID == 0)
                 {
-                    _orderItemRepository.DeleteOrderItem(Convert.ToInt64(id));
+                    orderParam.OrderID = _orderRepository.AddOrder(_mapper.Map<Demo.Core.Data.Model.Order>(orderParam));
+                }
+                else
+                {
+                     _orderRepository.UpdateOrder(_mapper.Map<Demo.Core.Data.Model.Order>(orderParam));
+                }
+
+                //OrderItems table
+                foreach (var item in orderParam.OrderItems)
+                {
+                    if (item.OrderItemID == 0)
+                    {
+                        item.OrderID = orderParam.OrderID;
+                        _orderItemRepository.AddOrderItem(_mapper.Map<Demo.Core.Data.Model.OrderItem>(item));
+                    }
+                    else
+                    {
+                        item.OrderID = orderParam.OrderID;
+                        _orderItemRepository.UpdateOrderItem(_mapper.Map<Demo.Core.Data.Model.OrderItem>(item));
+                    }
+                }
+                // Delete for OrderItems
+                if (orderParam.DeletedOrderItemIDs != null && orderParam.DeletedOrderItemIDs != "")
+                {
+                    foreach (var id in orderParam.DeletedOrderItemIDs.Split(',').Where(x => x != ""))
+                    {
+                        _orderItemRepository.DeleteOrderItem(Convert.ToInt64(id));
+                    }
                 }
                
-                return Ok();
+                return Ok(200);
             }
             catch (Exception ex)
             {
@@ -59,16 +72,23 @@ namespace DemoWebApiService.Controllers
         }
 
         // DELETE: api/Order/5
-        [HttpPost("api/Order/DeleteOrder/{id}")]
+        [HttpDelete("api/Order/DeleteOrder/{id}")]
         public IActionResult DeleteOrder(long id)
         {
-            var _orderItems = _mapper.Map<IEnumerable<OrderItem>>(_orderItemRepository.GetOrderItemsByOrderID(id));
-            foreach (var item in _orderItems.ToList())
+            try
             {
-                _orderItemRepository.DeleteOrderItem(item.OrderItemID);
+                var _orderItems = _mapper.Map<IEnumerable<OrderItem>>(_orderItemRepository.GetOrderItemsByOrderID(id));
+                foreach (var item in _orderItems.ToList())
+                {
+                    _orderItemRepository.DeleteOrderItem(item.OrderItemID);
+                }
+                _orderRepository.DeleteOrder(id);
+                return Ok();
             }
-            _orderRepository.DeleteOrder(id);
-            return Ok();
+            catch (Exception ex)
+            {
+                throw ex;
+            }
         }
 
         [HttpGet("api/Order/GetAllOrders")]
@@ -76,22 +96,26 @@ namespace DemoWebApiService.Controllers
         {
             try
             {
-                var _order = _mapper.Map<IEnumerable<Order>>(_orderRepository.GetOrders());
-                return _order == null ? NotFound(_order) : (IActionResult)Ok(_order);
+                OrderList obj = new OrderList();
+                var _order = _orderRepository.GetOrders();
+                 obj.OrderLists = _mapper.Map<IEnumerable<Order>>(_order);
+                obj.TotalCount = _order.Count();
+                return Ok(obj);
             }
             catch (Exception ex)
             {
                 throw ex;
             }
-}
-        [HttpGet("api/Order/GetOrders/{Id}")]
+        }
+        [HttpGet("api/Order/GetOrderAndOrderItem/{Id}")]
         public IActionResult GetOrderAndOrderItem(long id)
         {
             try
             {
-                var _order = _mapper.Map<Order>(_orderRepository.GetOrderByID(id));
-                var _orderItems = _mapper.Map<IEnumerable<OrderItem>>(_orderItemRepository.GetOrderItemsByOrderID(id));
-                return Ok(new { _order, _orderItems });
+
+                Order _order = _mapper.Map<Order>(_orderRepository.GetOrderByID(id));
+                _order.OrderItems = _mapper.Map<IEnumerable<OrderItem>>(_orderItemRepository.GetOrderItemsByOrderID(id));
+                return Ok(_order);
             }
             catch (Exception ex)
             {
